#68: Enabled/disabled states for widgets => closes #68

+++ Source/Gorgon/Widgets/Inputbox.cpp
@@ -230,7 +230,7 @@ namespace Gorgon { namespace Widgets { namespace internal {
                     
                     display.insert(selstart.byte, s);
                     
-                    selstart.byte  += s.size();
+                    selstart.byte  += (int)s.size();
                     selstart.glyph += gcnt;
                     
                     glyphcount += gcnt;
